requirements.txt
# general
mypy
pylint
autopep8==2.0.1
# others
Django>=4.0.3
mongoengine>=0.23.1
pymongo>=3.12.3
django-environ>=0.8.1
colorlog>=6.6.0
djangorestframework>=3.12.4
selenium>=4.1.3
spacy>=3.3.0
langdetect>=1.0.9
translate-api>=4.9.5
ru-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.3.0/ru_core_news_sm-3.3.0-py3-none-any.whl
zh-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/zh_core_web_sm-3.3.0/zh_core_web_sm-3.3.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
sqlalchemy==1.4.37
deep-translator==1.9.1
pandas
# tests
pytest-cov==3.0.0
pytest-django==4.5.2
pytest==7.1.2
pytest-timeout==2.1.0